Se agregan modulos exercise, routine_exercise,user_workout_routine

+++ back/src/exercise/exercise.service.ts
@@ -30,6 +30,10 @@ export class ExerciseService {
   }
 
   async delete(id: string): Promise<void> {
+    const exists = await this.exerciseRepository.exists(id);
+    if (!exists) {
+      throw new NotFoundException('Exercise not found');
+    }
     await this.exerciseRepository.delete(id);
   }
 }
